ejercicio_entregable_4/requirements.txt
pandas
django
requests
djangorestframework
mysqlclient